pyproject.toml
[tool.poetry]
name = "llm-agent"
version = "0.1.0"
description = ""
authors = ["Jiři Filip <wonttell@example.com>"]
readme = "README.md"
packages = [
    {include = "llm_agent", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.1.0"
pydantic = "^2.10.6"
anthropic = "^0.45.2"


[tool.poetry.group.research.dependencies]
jupyterlab = "^4.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
